+++ packages/enhanced/src/lib/container/ContainerEntryDependency.ts
@@ -6,8 +6,6 @@
 import { ExposeOptions } from './ContainerEntryModule';
 import { normalizeWebpackPath } from '@module-federation/sdk/normalize-webpack-path';
 
-import type { containerPlugin } from '@module-federation/sdk';
-
 const makeSerializable = require(
   normalizeWebpackPath('webpack/lib/util/makeSerializable'),
 );
@@ -20,28 +18,24 @@ class ContainerEntryDependency extends Dependency {
   public exposes: [string, ExposeOptions][];
   public shareScope: string | string[];
   public injectRuntimeEntry: string;
-  public dataPrefetch: containerPlugin.ContainerPluginOptions['dataPrefetch'];
 
   /**
    * @param {string} name entry name
    * @param {[string, ExposeOptions][]} exposes list of exposed modules
    * @param {string|string[]} shareScope name of the share scope
    * @param {string[]} injectRuntimeEntry the path of injectRuntime file.
-   * @param {containerPlugin.ContainerPluginOptions['dataPrefetch']} dataPrefetch whether enable dataPrefetch
    */
   constructor(
     name: string,
     exposes: [string, ExposeOptions][],
     shareScope: string | string[],
     injectRuntimeEntry: string,
-    dataPrefetch: containerPlugin.ContainerPluginOptions['dataPrefetch'],
   ) {
     super();
     this.name = name;
     this.exposes = exposes;
     this.shareScope = shareScope;
     this.injectRuntimeEntry = injectRuntimeEntry;
-    this.dataPrefetch = dataPrefetch;
   }
 
   /**