+++ packages/runtime-core/src/utils/preload.ts
@@ -20,7 +20,6 @@ export function defaultPreloadArgs(
     resourceCategory: 'sync',
     share: true,
     depsRemote: true,
-    prefetchInterface: false,
     ...preloadConfig,
   } as PreloadConfig;
 }