+++ packages/runtime/src/index.ts
@@ -26,7 +26,10 @@ export function createInstance(options: UserOptions) {
   // Retrieve debug constructor
   const ModuleFederationConstructor =
     getGlobalFederationConstructor() || ModuleFederation;
-  const instance = new ModuleFederationConstructor(options);
+  const instance = new ModuleFederationConstructor({
+    id: `${options.name}@${options.version || Date.now()}`,
+    ...options,
+  });
   setGlobalFederationInstance(instance);
   return instance;
 }
@@ -38,12 +41,13 @@ let FederationInstance: ModuleFederation | null = null;
 export function init(options: UserOptions): ModuleFederation {
   // Retrieve the same instance with the same name
   const instance = getGlobalFederationInstance(options.name, options.version);
+  const normalizedOptions = { ...options, id: options.id || '' };
   if (!instance) {
-    FederationInstance = createInstance(options);
+    FederationInstance = createInstance(normalizedOptions);
     return FederationInstance;
   } else {
     // Merge options
-    instance.initOptions(options);
+    instance.initOptions(normalizedOptions);
     if (!FederationInstance) {
       FederationInstance = instance;
     }