+++ packages/sdk/src/types/stats.ts
@@ -30,6 +30,7 @@ export interface ResourceInfo {
 export interface StatsBuildInfo {
   buildVersion: string;
   buildName: string;
+  hash?: string;
 
   // only appear when enable treeshaking
   target?: string[];
@@ -50,8 +51,6 @@ export interface BasicStatsMetaData {
   buildInfo: StatsBuildInfo;
   remoteEntry: ResourceInfo;
   ssrRemoteEntry?: ResourceInfo;
-  prefetchInterface?: boolean;
-  prefetchEntry?: ResourceInfo;
   types?: MetaDataTypes;
   type: string;
   pluginVersion?: string;
@@ -129,6 +128,7 @@ export interface StatsExpose {
   file: string;
   requires: string[];
   assets: StatsAssets;
+  hash?: string;
 }
 
 export interface Stats<T = BasicStatsMetaData, K = StatsRemoteVal> {