+++ apps/node-dynamic-remote/CHANGELOG.md
@@ -1,5 +1,11 @@
 # node-dynamic-remote
 
+## 1.0.9
+
+### Patch Changes
+
+- @module-federation/node@2.7.42
+
 ## 1.0.8
 
 ### Patch Changes