+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/cli
 
+## 2.4.0
+
+### Patch Changes
+
+- Updated dependencies [5eba770]
+- Updated dependencies [13b1e84]
+  - @module-federation/sdk@2.4.0
+  - @module-federation/dts-plugin@2.4.0
+
 ## 2.3.3
 
 ### Patch Changes