+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @module-federation/nextjs-mf
 
+## 8.8.66
+
+### Patch Changes
+
+- Updated dependencies [5eba770]
+- Updated dependencies [3a3414c]
+- Updated dependencies [13b1e84]
+  - @module-federation/sdk@2.4.0
+  - @module-federation/runtime@2.4.0
+  - @module-federation/enhanced@2.4.0
+  - @module-federation/webpack-bundler-runtime@2.4.0
+  - @module-federation/node@2.7.42
+
 ## 8.8.65
 
 ### Patch Changes