+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/rsbuild-plugin
 
+## 2.4.0
+
+### Patch Changes
+
+- Updated dependencies [5eba770]
+- Updated dependencies [13b1e84]
+  - @module-federation/sdk@2.4.0
+  - @module-federation/enhanced@2.4.0
+  - @module-federation/node@2.7.42
+
 ## 2.3.3
 
 ### Patch Changes