+++ packages/runtime-plugins/inject-external-runtime-core-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/inject-external-runtime-core-plugin
 
+## 2.4.0
+
+### Patch Changes
+
+- @module-federation/runtime-tools@2.4.0
+
 ## 2.3.3
 
 ### Patch Changes