+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @module-federation/third-party-dts-extractor
 
+## 2.4.0
+
 ## 2.3.3
 
 ## 2.3.2